third_party/WebKit/LayoutTests/platform/mac-mac10.11/external/wpt/XMLHttpRequest/getresponseheader-chunked-trailer-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: getResponseHeader() and HTTP trailer assert_regexp_match: expected object "/trailer:\sX-Test-Me/" but got "Date: Fri, 14 Apr 2017 23:24:56 GMT\r\nTrailer: X-Test-Me\r\nServer: BaseHTTP/0.3 Python/2.7.10\r\nTransfer-Encoding: chunked\r\nContent-Type: text/plain\r\n"
Harness: the test ran to completion.
